auto_committer/requirements.txt
gitpython==3.1.43
watchdog==6.0.0
